backend/requirements.txt
# ─── Web framework ────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
websockets==12.0

# ─── LLM (CPU + optional GPU) ─────────────────────────────
llama-cpp-python==0.2.77      # install with CMAKE_ARGS="-DLLAMA_CUBLAS=on" for GPU

# ─── STT ──────────────────────────────────────────────────
faster-whisper==1.0.2

# ─── TTS ──────────────────────────────────────────────────
piper-tts==1.2.0

# ─── VAD ──────────────────────────────────────────────────
silero-vad==5.1
torch==2.2.2                  # CPU build — no CUDA needed for VAD
torchaudio==2.2.2

# ─── Audio processing ─────────────────────────────────────
numpy==1.26.4
soundfile==0.12.1
librosa==0.10.1

# ─── Resume parsing ───────────────────────────────────────
pymupdf==1.24.3               # PDF parsing (fitz)
python-docx==1.1.0            # DOCX parsing

# ─── Utilities ────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
aiofiles==23.2.1
httpx==0.27.0
